+++ service/src/test/kotlin/fi/espoo/evaka/invoicing/service/FeeDecisionGeneratorV2Test.kt
@@ -221,6 +221,7 @@ private fun createChildFeeBasis(
                 hasServiceNeed = true,
                 serviceNeedOption = snDaycareFullDay35
             ),
+        serviceNeedOptionFee = null,
         feeAlterations = emptyList(),
         income = null
     )
